refactor(jans-cedarling): split `mod.rs` into `error.rs`, `service.rs` for better organization
- Extracted error handling logic into `error.rs`
- Moved service-related functionality into `service.rs`
- Improved modularity and readability of the JWT module by reorganizing code

This refactor improves maintainability by splitting the large `mod.rs` into smaller, focused files.
